Moved data linking functions to linking folder
